+++ src/adapters/curve/common/getPoolData.ts
@@ -3,7 +3,7 @@ import { Erc20__factory } from '../../../contracts'
 import { Chain } from '../../../core/constants/chains'
 import { ZERO_ADDRESS } from '../../../core/constants/ZERO_ADDRESS'
 import { getTokenMetadata } from '../../../core/utils/getTokenMetadata'
-import { CustomJsonRpcProvider } from '../../../core/utils/provider/CustomJsonRpcProvider'
+import { CustomJsonRpcProvider } from '../../../core/provider/CustomJsonRpcProvider'
 import { Erc20Metadata } from '../../../types/erc20Metadata'
 import { MetaRegistry__factory } from '../contracts'
 import { CURVE_META_REGISTRY_CONTRACT } from '../products/pool/curvePoolAdapter'
