+++ src/adapters/iziswap/products/iziswap/iZiswapAdapter.ts
@@ -6,7 +6,7 @@ import { ResolveUnderlyingPositions } from '../../../../core/decorators/resolveU
 import { NotImplementedError } from '../../../../core/errors/errors'
 import { filterMapAsync } from '../../../../core/utils/filters'
 import { getTokenMetadata } from '../../../../core/utils/getTokenMetadata'
-import { CustomJsonRpcProvider } from '../../../../core/utils/provider/CustomJsonRpcProvider'
+import { CustomJsonRpcProvider } from '../../../../core/provider/CustomJsonRpcProvider'
 import {
   ProtocolAdapterParams,
   ProtocolDetails,
