+++ src/core/provider/ChainProvider.ts
@@ -1,9 +1,9 @@
 import { Network } from 'ethers'
-import { IConfig } from '../../../config'
-import { Multicall__factory } from '../../../contracts'
-import { Chain } from '../../constants/chains'
-import { MULTICALL_ADDRESS } from '../../constants/MULTICALL_ADDRESS'
-import { logger } from '../logger'
+import { IConfig } from '../../config'
+import { Multicall__factory } from '../../contracts'
+import { Chain } from '../constants/chains'
+import { MULTICALL_ADDRESS } from '../constants/MULTICALL_ADDRESS'
+import { logger } from '../utils/logger'
 import { CustomJsonRpcProvider } from './CustomJsonRpcProvider'
 import { CustomMulticallJsonRpcProvider } from './CustomMulticallJsonRpcProvider'
 import { MulticallQueue } from './MulticallQueue'
