+++ src/defiProvider.ts
@@ -8,9 +8,9 @@ import { Chain, ChainName } from './core/constants/chains'
 import { TimePeriod } from './core/constants/timePeriod'
 import { ProviderMissingError } from './core/errors/errors'
 import { getProfits } from './core/getProfits'
+import { ChainProvider } from './core/provider/ChainProvider'
+import { CustomJsonRpcProvider } from './core/provider/CustomJsonRpcProvider'
 import { logger } from './core/utils/logger'
-import { ChainProvider } from './core/utils/provider/ChainProvider'
-import { CustomJsonRpcProvider } from './core/utils/provider/CustomJsonRpcProvider'
 import {
   enrichPositionBalance,
   enrichUnderlyingTokenRates,
