+++ src/scripts/buildMetadata.ts
@@ -11,7 +11,7 @@ import { IMetadataBuilder } from '../core/decorators/cacheToFile'
 import { ProviderMissingError } from '../core/errors/errors'
 import { pascalCase } from '../core/utils/caseConversion'
 import { logger } from '../core/utils/logger'
-import { CustomJsonRpcProvider } from '../core/utils/provider/CustomJsonRpcProvider'
+import { CustomJsonRpcProvider } from '../core/provider/CustomJsonRpcProvider'
 import { writeCodeFile } from '../core/utils/writeCodeFile'
 import { Json } from '../types/json'
 import { multiChainFilter, multiProtocolFilter } from './commandFilters'
