+++ src/scripts/buildSnapshots.ts
@@ -6,7 +6,7 @@ import { Chain, ChainName } from '../core/constants/chains'
 import { ProviderMissingError } from '../core/errors/errors'
 import { bigintJsonStringify } from '../core/utils/bigintJson'
 import { kebabCase } from '../core/utils/caseConversion'
-import { CustomJsonRpcProvider } from '../core/utils/provider/CustomJsonRpcProvider'
+import { CustomJsonRpcProvider } from '../core/provider/CustomJsonRpcProvider'
 import { DefiProvider } from '../defiProvider'
 import type { TestCase } from '../types/testCase'
 import { multiProtocolFilter } from './commandFilters'
